+++ app/login/page.tsx
@@ -1,5 +1,5 @@
 import { headers, cookies } from 'next/headers';
-import { createClient } from '@/utils/supabase/server';
+import { createClient } from '@/lib/supabase/server';
 import { redirect } from 'next/navigation';
 import Link from 'next/link';
