+++ middleware.ts
@@ -1,4 +1,4 @@
-import { createClient } from '@/utils/supabase/middleware';
+import { createClient } from '@/lib/supabase/middleware';
 import type { NextRequest } from 'next/server';
 
 export async function middleware(request: NextRequest) {
